# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AccessibilityResponse": {
        "additionalProperties": false,
        "description": "Accessibility advice calculated for one exact immutable Revision.",
        "properties": {
          "revisionId": {
            "format": "uuid",
            "title": "Revisionid",
            "type": "string"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/AccessibilityWarningResponse"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "revisionId",
          "warnings"
        ],
        "title": "AccessibilityResponse",
        "type": "object"
      },
...
      "CommentResolutionRequest": {
        "additionalProperties": false,
        "description": "The next derived resolution state for one immutable comment.",
        "properties": {
          "resolved": {
            "title": "Resolved",
            "type": "boolean"
          }
        },
        "required": [
          "resolved"
        ],
        "title": "CommentResolutionRequest",
        "type": "object"
      },
...
      "EditResponse": {
        "additionalProperties": false,
        "description": "One Edit and the composition its current Revision holds.",
        "properties": {
          "brandViolations": {
            "items": {
              "$ref": "#/components/schemas/BrandViolationResponse"
            },
            "title": "Brandviolations",
            "type": "array"
          },
          "candidateId": {
            "format": "uuid",
            "title": "Candidateid",
            "type": "string"
          },
          "composition": {
            "$ref": "#/components/schemas/CompositionV1"
          },
          "compositionHash": {
            "title": "Compositionhash",
            "type": "string"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "currentRevision": {
            "title": "Currentrevision",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "projectId": {
            "format": "uuid",
            "title": "Projectid",
            "type": "string"
          },
          "updatedAt": {
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "candidateId",
          "currentRevision",
          "composition",
          "compositionHash",
          "brandViolations",
          "createdAt",
          "updatedAt"
        ],
        "title": "EditResponse",
        "type": "object"
      },
      "EditReviewDecisionKind": {
        "description": "The decisions a reviewer may record against one immutable Revision.",
        "enum": [
          "request_changes",
          "approve"
        ],
        "title": "EditReviewDecisionKind",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Platform": {
        "description": "The short-form destinations Clipah packages for, without publishing to any of them.",
        "enum": [
          "tiktok",
          "instagram_reels",
          "youtube_shorts"
        ],
        "title": "Platform",
        "type": "string"
      },
...
      "ReviewCommentRequest": {
        "additionalProperties": false,
        "description": "One plain-text comment against an exact immutable Revision.",
        "properties": {
          "anchor": {
            "$ref": "#/components/schemas/ReviewAnchorRequest"
          },
          "revisionId": {
            "format": "uuid",
            "title": "Revisionid",
            "type": "string"
          },
          "text": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "revisionId",
          "text",
          "anchor"
        ],
        "title": "ReviewCommentRequest",
        "type": "object"
      },
      "ReviewCommentResponse": {
        "additionalProperties": false,
        "description": "One review comment with its derived resolution state.",
        "properties": {
          "actorDisplayName": {
            "title": "Actordisplayname",
            "type": "string"
          },
          "actorUserId": {
            "format": "uuid",
            "title": "Actoruserid",
            "type": "string"
          },
          "anchor": {
            "$ref": "#/components/schemas/ReviewAnchorResponse"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "resolved": {
            "title": "Resolved",
            "type": "boolean"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "revisionId": {
            "format": "uuid",
            "title": "Revisionid",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "revisionId",
          "revision",
          "actorUserId",
          "actorDisplayName",
          "text",
          "anchor",
          "resolved",
          "createdAt"
        ],
        "title": "ReviewCommentResponse",
        "type": "object"
      },
      "ReviewDecisionRequest": {
        "additionalProperties": false,
        "description": "One approval or request for changes against exact bytes.",
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/EditReviewDecisionKind"
          },
          "revisionId": {
            "format": "uuid",
            "title": "Revisionid",
            "type": "string"
          }
        },
        "required": [
          "revisionId",
          "decision"
        ],
        "title": "ReviewDecisionRequest",
        "type": "object"
      },
      "ReviewDecisionResponse": {
        "additionalProperties": false,
        "description": "One historical decision and whether it still targets current bytes.",
        "properties": {
          "actorDisplayName": {
            "title": "Actordisplayname",
            "type": "string"
          },
          "actorUserId": {
            "format": "uuid",
            "title": "Actoruserid",
            "type": "string"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "current": {
            "title": "Current",
            "type": "boolean"
          },
          "decision": {
            "$ref": "#/components/schemas/EditReviewDecisionKind"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "revisionId": {
            "format": "uuid",
            "title": "Revisionid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "revisionId",
          "revision",
          "actorUserId",
          "actorDisplayName",
          "decision",
          "current",
          "createdAt"
        ],
        "title": "ReviewDecisionResponse",
        "type": "object"
      },
      "ReviewSummaryResponse": {
        "additionalProperties": false,
        "description": "The current review state and immutable history for one Edit.",
        "properties": {
          "approved": {
            "title": "Approved",
            "type": "boolean"
          },
          "comments": {
            "items": {
              "$ref": "#/components/schemas/ReviewCommentResponse"
            },
            "title": "Comments",
            "type": "array"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/ReviewDecisionResponse"
            },
            "title": "Decisions",
            "type": "array"
          }
        },
        "required": [
          "approved",
          "comments",
          "decisions"
        ],
        "title": "ReviewSummaryResponse",
        "type": "object"
      },
...
      "SaveRevisionRequest": {
        "additionalProperties": false,
        "description": "One composition offered as the successor of the Revision the caller holds.",
        "properties": {
          "composition": {
            "additionalProperties": true,
            "title": "Composition",
            "type": "object"
          },
          "expectedRevision": {
            "minimum": 1.0,
            "title": "Expectedrevision",
            "type": "integer"
          }
        },
        "required": [
          "expectedRevision",
          "composition"
        ],
        "title": "SaveRevisionRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Clipah API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/edit-review-comments/{comment_id}/resolution": {
      "post": {
        "description": "Append one resolution or reopening event for a review comment.",
        "operationId": "update_comment_resolution_api_v1_edit_review_comments__comment_id__resolution_post",
        "parameters": [
          {
            "in": "path",
            "name": "comment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Comment Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentResolutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewCommentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Comment Resolution",
        "tags": [
          "edit-reviews"
        ]
      }
    },
    "/api/v1/edits/{edit_id}": {
      "get": {
        "description": "Read one Edit, hiding another Workspace's Edit behind the same absence.",
        "operationId": "show_api_v1_edits__edit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "edit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Edit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Show",
        "tags": [
          "edits"
        ]
      },
      "put": {
        "description": "Append one composition as the next Revision, or refuse a stale expectation.",
        "operationId": "save_api_v1_edits__edit_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "edit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Edit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save",
        "tags": [
          "edits"
        ]
      }
    },
    "/api/v1/edits/{edit_id}/accessibility": {
      "get": {
        "description": "Measure one immutable Revision without changing its composition.",
        "operationId": "show_accessibility_quality_api_v1_edits__edit_id__accessibility_get",
        "parameters": [
          {
            "in": "path",
            "name": "edit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Edit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Platform"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessibilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Show Accessibility Quality",
        "tags": [
          "edit-reviews"
        ]
      }
    },
...
    "/api/v1/edits/{edit_id}/review-comments": {
      "post": {
        "description": "Append one comment to an immutable Revision.",
        "operationId": "create_review_comment_api_v1_edits__edit_id__review_comments_post",
        "parameters": [
          {
            "in": "path",
            "name": "edit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Edit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewCommentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Review Comment",
        "tags": [
          "edit-reviews"
        ]
      }
    },
    "/api/v1/edits/{edit_id}/reviews": {
      "get": {
        "description": "Read the current answer and immutable review history.",
        "operationId": "show_edit_review_api_v1_edits__edit_id__reviews_get",
        "parameters": [
          {
            "in": "path",
            "name": "edit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Edit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Show Edit Review",
        "tags": [
          "edit-reviews"
        ]
      },
      "post": {
        "description": "Append one approval or request for changes.",
        "operationId": "create_review_decision_api_v1_edits__edit_id__reviews_post",
        "parameters": [
          {
            "in": "path",
            "name": "edit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Edit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Review Decision",
        "tags": [
          "edit-reviews"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/candidates/{candidate_id}/edits": {
      "post": {
        "description": "Open the one Edit belonging to a reviewed candidate, or reach the existing one.",
        "operationId": "create_api_v1_projects__project_id__candidates__candidate_id__edits_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/EditSelectionRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Selection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "edits"
        ]
      }
    },
...
  }
}